pyproject.toml
[project]
name = "ikf-ai-skills"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.3.22",
    "fastapi>=0.128.0",
    "google-genai>=1.56.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-fasthtml>=0.12.37",
    "sqlalchemy>=2.0.45",
    "tavily-python>=0.7.17",
    "uvicorn>=0.40.0",
]
